% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_reconstruction_uncertainty}
\alias{add_reconstruction_uncertainty}
\title{Add Reconstruction Uncertainty}
\usage{
add_reconstruction_uncertainty(stan_data, network, recon_opts)
}
\description{
Add Reconstruction Uncertainty
}
\keyword{internal}
