% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_tau}
\alias{diagnose_tau}
\title{Diagnose Tau Selection}
\usage{
diagnose_tau(network, max_tau = NULL, grid_size = 20)
}
\arguments{
\item{network}{An rmst_network object}

\item{max_tau}{Maximum tau to consider}

\item{grid_size}{Number of tau values to evaluate}
}
\value{
A list with tau diagnostics
}
\description{
Provides diagnostics for selecting appropriate tau values for RMST analysis
}
