% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{export_cinema}
\alias{export_cinema}
\title{Export to CINeMA Format}
\usage{
export_cinema(fit, file, tau = NULL, ...)
}
\arguments{
\item{fit}{An rmst_nma_fit object}

\item{file}{Output filename}

\item{tau}{Specific tau value to export}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the exported data frame
}
\description{
Exports RMST NMA results to CINeMA-compatible CSV format
}
