% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_validation.R
\name{list_nma_datasets}
\alias{list_nma_datasets}
\title{List available packaged NMA datasets}
\usage{
list_nma_datasets()
}
\value{
A character vector of dataset object names
}
\description{
Lists \code{.rda} datasets shipped in the package \verb{data/} directory that match
the expected format.
}
