% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_validation.R
\name{nma_schema}
\alias{nma_schema}
\title{Return the canonical NMA dataset schema}
\usage{
nma_schema()
}
\value{
A list with \code{required} and \code{optional} named vectors of types
}
\description{
Return the canonical NMA dataset schema
}
