% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.rmst_nma_fit}
\alias{plot.rmst_nma_fit}
\title{Plot RMST NMA Results}
\usage{
\method{plot}{rmst_nma_fit}(
  x,
  type = c("forest", "rank", "rmst", "trace", "network"),
  tau = NULL,
  ...
)
}
\arguments{
\item{x}{An rmst_nma_fit object}

\item{type}{Type of plot ("forest", "rank", "rmst", "trace", "network")}

\item{tau}{Specific tau value for plotting (if multiple fitted)}

\item{...}{Additional plotting arguments}
}
\value{
A ggplot object
}
\description{
Creates various plots for RMST NMA results
}
