% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rankogram}
\alias{plot_rankogram}
\title{Rankogram plot using Stan's generated quantities}
\usage{
plot_rankogram(fit, tau, ...)
}
\description{
Rankogram plot using Stan's generated quantities
}
\keyword{internal}
