% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst_nma.R
\name{prepare_stan_data}
\alias{prepare_stan_data}
\title{Prepare Stan data}
\usage{
prepare_stan_data(network, tau, baseline, random_effects, inconsistency)
}
\description{
Prepare Stan data
}
\keyword{internal}
