% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst_network.R
\name{print.rmst_network}
\alias{print.rmst_network}
\title{Print method for rmst_network}
\usage{
\method{print}{rmst_network}(x, ...)
}
\arguments{
\item{x}{An rmst_network object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for rmst_network
}
