% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recon_opts}
\alias{recon_opts}
\title{Reconstruction Options}
\usage{
recon_opts(R = 50, method = c("envelope", "bootstrap"), cache = TRUE)
}
\arguments{
\item{R}{Number of reconstruction draws}

\item{method}{Reconstruction method}

\item{cache}{Whether to cache results}
}
\value{
List of reconstruction options
}
\description{
Set options for KM reconstruction uncertainty propagation
}
