% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst_nma.R
\name{rmst_nma}
\alias{rmst_nma}
\title{#' Fit RMST Network Meta-Analysis}
\usage{
rmst_nma(
  network,
  tau = 12,
  baseline = "rp_spline",
  random_effects = "normal",
  inconsistency = FALSE,
  propagate_recon_uncertainty = FALSE,
  recon = NULL,
  chains = 4,
  iter = 2000,
  warmup = NULL,
  adapt_delta = 0.9,
  ...
)
}
\arguments{
\item{network}{An rmst_network object}

\item{tau}{Restriction time(s) for RMST calculation}

\item{baseline}{Baseline hazard model ("rp_spline" or "piecewise")}

\item{random_effects}{Random effects distribution ("normal" or "student_t")}

\item{inconsistency}{Include inconsistency parameters}

\item{propagate_recon_uncertainty}{Propagate reconstruction uncertainty}

\item{recon}{Reconstruction options}

\item{chains}{Number of MCMC chains}

\item{iter}{Number of iterations per chain}

\item{warmup}{Number of warmup iterations}

\item{adapt_delta}{Target acceptance probability}

\item{...}{Additional arguments to cmdstanr}
}
\value{
An object of class 'rmst_nma_fit'
}
\description{
Fits a Bayesian network meta-analysis model for RMST outcomes
}
