% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.rmst_nma_fit}
\alias{summary.rmst_nma_fit}
\title{Summary method for rmst_nma_fit}
\usage{
\method{summary}{rmst_nma_fit}(object, ...)
}
\arguments{
\item{object}{An rmst_nma_fit object}

\item{...}{Additional arguments}
}
\value{
Summary object
}
\description{
Summary method for rmst_nma_fit
}
